remove example mod codes
