--- common/src/main/java/net/examplemod/ExampleMod.java
@@ -1,33 +1,0 @@
-package net.examplemod;
-
-import me.shedaniel.architectury.registry.CreativeTabs;
-import me.shedaniel.architectury.registry.DeferredRegister;
-import me.shedaniel.architectury.registry.Registries;
-import me.shedaniel.architectury.registry.RegistrySupplier;
-import net.minecraft.core.Registry;
-import net.minecraft.resources.ResourceLocation;
-import net.minecraft.util.LazyLoadedValue;
-import net.minecraft.world.item.CreativeModeTab;
-import net.minecraft.world.item.Item;
-import net.minecraft.world.item.ItemStack;
-
-import java.util.function.Supplier;
-
-public class ExampleMod {
-    public static final String MOD_ID = "examplemod";
-    // We can use this if we don't want to use DeferredRegister
-    public static final LazyLoadedValue<Registries> REGISTRIES = new LazyLoadedValue<>(() -> Registries.get(MOD_ID));
-    // Registering a new creative tab
-    public static final CreativeModeTab EXAMPLE_TAB = CreativeTabs.create(new ResourceLocation(MOD_ID, "example_tab"), () ->
-            new ItemStack(ExampleMod.EXAMPLE_ITEM.get()));
-    
-    public static final DeferredRegister<Item> ITEMS = DeferredRegister.create(MOD_ID, Registry.ITEM_REGISTRY);
-    public static final RegistrySupplier<Item> EXAMPLE_ITEM = ITEMS.register("example_item", () ->
-            new Item(new Item.Properties().tab(ExampleMod.EXAMPLE_TAB)));
-    
-    public static void init() {
-        ITEMS.register();
-        
-        System.out.println(ExampleExpectPlatform.getConfigDirectory().toAbsolutePath().normalize().toString());
-    }
-}
