--- fabric/src/main/java/net/examplemod/fabric/ExampleModFabric.java
@@ -1,11 +1,0 @@
-package net.examplemod.fabric;
-
-import net.examplemod.ExampleMod;
-import net.fabricmc.api.ModInitializer;
-
-public class ExampleModFabric implements ModInitializer {
-    @Override
-    public void onInitialize() {
-        ExampleMod.init();
-    }
-}
