--- forge/src/main/java/net/examplemod/forge/ExampleExpectPlatformImpl.java
@@ -1,15 +1,0 @@
-package net.examplemod.forge;
-
-import net.examplemod.ExampleExpectPlatform;
-import net.minecraftforge.fml.loading.FMLPaths;
-
-import java.nio.file.Path;
-
-public class ExampleExpectPlatformImpl {
-    /**
-     * This is our actual method to {@link ExampleExpectPlatform#getConfigDirectory()}.
-     */
-    public static Path getConfigDirectory() {
-        return FMLPaths.CONFIGDIR.get();
-    }
-}
